+++ NEWS.md
@@ -1,6 +1,20 @@
+# anndataR 1.1.1
+
+- Fix CI (PR #418).
+- Add continuous benchmarking using bencher (PR #423, PR #425).
+- Handle unnamed `SingleCellExperiment` assays in `as_AnnData()` by automatically assigning names with a warning (PR #420).
+- Optimise sparse matrix reading performance by avoiding `Matrix::sparseMatrix` and constructing objects manually (PR #417).
+
 # anndataR 1.1.0
 
 - Bioconductor 3.23 devel
+
+# anndataR 1.0.1
+
+- Update author email address (PR #410).
+- Fix linting issues (PR #412).
+- Fix roundtrip test; ensure that `X` is always 2-dimensional (PR #413).
+- Fix HDF5 write test; make sure to only pass CsparseMatrix to Seurat (PR #413).
 
 # anndataR 1.0.0